refactor: add Time
